refactor: settings route
